+++ org.eclipse.xtext.xbase.ide/src/org/eclipse/xtext/xbase/annotations/ide/XbaseWithAnnotationsIdeModule.java
@@ -1,17 +1,14 @@
-/**
- * Copyright (c) 2010, 2017 itemis AG (http://www.itemis.eu) and others.
+/*******************************************************************************
+ * Copyright (c) 2010, 2020 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- */
+ *******************************************************************************/
 package org.eclipse.xtext.xbase.annotations.ide;
 
-import org.eclipse.xtext.xbase.annotations.ide.AbstractXbaseWithAnnotationsIdeModule;
-
-/**
+/** 
  * Use this class to register ide components.
  */
-@SuppressWarnings("all")
 public class XbaseWithAnnotationsIdeModule extends AbstractXbaseWithAnnotationsIdeModule {
 }
