--- org.eclipse.xtext.xbase.ide/src/org/eclipse/xtext/xbase/annotations/ide/XbaseWithAnnotationsIdeModule.xtend
@@ -1,14 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2010, 2017 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *******************************************************************************/
-package org.eclipse.xtext.xbase.annotations.ide
-
-/** 
- * Use this class to register ide components.
- */
-class XbaseWithAnnotationsIdeModule extends AbstractXbaseWithAnnotationsIdeModule {
-}
