--- org.eclipse.xtext.xbase.ide/xtend-gen/org/eclipse/xtext/xbase/ide/XbaseIdeModule.java
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) 2010, 2017 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.eclipse.xtext.xbase.ide;
-
-import org.eclipse.xtext.xbase.ide.AbstractXbaseIdeModule;
-
-/**
- * Use this class to register ide components.
- */
-@SuppressWarnings("all")
-public class XbaseIdeModule extends AbstractXbaseIdeModule {
-}
