+++ org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/XtypeRuntimeModule.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2010 itemis AG (http://www.itemis.eu) and others.
+/** 
+ * Copyright (c) 2010, 2020 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -7,12 +7,9 @@
  */
 package org.eclipse.xtext.xbase;
 
-import org.eclipse.xtext.xbase.AbstractXtypeRuntimeModule;
-
-/**
+/** 
  * Use this class to register components to be used within the IDE.
  * @since 2.7
  */
-@SuppressWarnings("all")
-public class XtypeRuntimeModule extends AbstractXtypeRuntimeModule {
+public class XtypeRuntimeModule extends org.eclipse.xtext.xbase.AbstractXtypeRuntimeModule {
 }
