--- org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/XtypeRuntimeModule.xtend
@@ -1,15 +1,0 @@
-/** 
- * Copyright (c) 2010 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.eclipse.xtext.xbase
-
-/** 
- * Use this class to register components to be used within the IDE.
- * @since 2.7
- */
-class XtypeRuntimeModule extends org.eclipse.xtext.xbase.AbstractXtypeRuntimeModule {
-}
