+++ org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/XtypeStandaloneSetup.java
@@ -1,4 +1,4 @@
-/**
+/** 
  * Copyright (c) 2010 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -7,27 +7,26 @@
  */
 package org.eclipse.xtext.xbase;
 
-import com.google.inject.Injector;
 import org.eclipse.emf.ecore.EPackage;
 import org.eclipse.xtext.common.types.TypesPackage;
-import org.eclipse.xtext.xbase.XtypeStandaloneSetupGenerated;
 import org.eclipse.xtext.xtype.XtypePackage;
+import com.google.inject.Injector;
 
-/**
- * Initialization support for running Xtext languages
+/** 
+ * Initialization support for running Xtext languages 
  * without equinox extension registry
  * @since 2.7
  */
-@SuppressWarnings("all")
 public class XtypeStandaloneSetup extends XtypeStandaloneSetupGenerated {
-  public static void doSetup() {
-    new XtypeStandaloneSetup().createInjectorAndDoEMFRegistration();
-  }
-  
-  @Override
-  public Injector createInjectorAndDoEMFRegistration() {
-    EPackage.Registry.INSTANCE.put(TypesPackage.eINSTANCE.getNsURI(), TypesPackage.eINSTANCE);
-    EPackage.Registry.INSTANCE.put(XtypePackage.eINSTANCE.getNsURI(), XtypePackage.eINSTANCE);
-    return super.createInjectorAndDoEMFRegistration();
-  }
+
+	public static void doSetup() {
+		new XtypeStandaloneSetup().createInjectorAndDoEMFRegistration();
+	}
+
+	@Override
+	public Injector createInjectorAndDoEMFRegistration() {
+		EPackage.Registry.INSTANCE.put(TypesPackage.eINSTANCE.getNsURI(), TypesPackage.eINSTANCE);
+		EPackage.Registry.INSTANCE.put(XtypePackage.eINSTANCE.getNsURI(), XtypePackage.eINSTANCE);
+		return super.createInjectorAndDoEMFRegistration();
+	}
 }
