+++ org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/annotations/XbaseWithAnnotationsRuntimeModule.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2011 itemis AG (http://www.itemis.eu) and others.
+/** 
+ * Copyright (c) 2011, 2020 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -7,11 +7,8 @@
  */
 package org.eclipse.xtext.xbase.annotations;
 
-import org.eclipse.xtext.xbase.annotations.AbstractXbaseWithAnnotationsRuntimeModule;
-
-/**
+/** 
  * Use this class to register components to be used at runtime / without the Equinox extension registry.
  */
-@SuppressWarnings("all")
-public class XbaseWithAnnotationsRuntimeModule extends AbstractXbaseWithAnnotationsRuntimeModule {
+public class XbaseWithAnnotationsRuntimeModule extends org.eclipse.xtext.xbase.annotations.AbstractXbaseWithAnnotationsRuntimeModule {
 }
