--- org.eclipse.xtext.xbase/src/org/eclipse/xtext/xbase/annotations/XbaseWithAnnotationsRuntimeModule.xtend
@@ -1,14 +1,0 @@
-/** 
- * Copyright (c) 2011 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.eclipse.xtext.xbase.annotations
-
-/** 
- * Use this class to register components to be used at runtime / without the Equinox extension registry.
- */
-class XbaseWithAnnotationsRuntimeModule extends org.eclipse.xtext.xbase.annotations.AbstractXbaseWithAnnotationsRuntimeModule {
-}
